AP_Radio: switch to 16 bit TX version
using year, month, day

+++ libraries/AP_Radio/AP_Radio.cpp
@@ -167,7 +167,7 @@ void AP_Radio::update(void)
 }
 
 // get transmitter firmware version
-uint8_t AP_Radio::get_tx_version(void)
+uint32_t AP_Radio::get_tx_version(void)
 {
     if (driver) {
         return driver->get_tx_version();
